Stop adding to server.go...
